+++ src/main/java/com/idea5/four_cut_photos_map/domain/memberTitle/service/MemberTitleService.java
@@ -1,15 +1,12 @@
 package com.idea5.four_cut_photos_map.domain.memberTitle.service;
 
-import com.idea5.four_cut_photos_map.domain.favorite.service.FavoriteService;
 import com.idea5.four_cut_photos_map.domain.member.entity.Member;
 import com.idea5.four_cut_photos_map.domain.memberTitle.dto.response.MemberTitleInfoResp;
 import com.idea5.four_cut_photos_map.domain.memberTitle.dto.response.MemberTitleResp;
 import com.idea5.four_cut_photos_map.domain.memberTitle.entity.MemberTitle;
 import com.idea5.four_cut_photos_map.domain.memberTitle.entity.MemberTitleLog;
-import com.idea5.four_cut_photos_map.domain.memberTitle.entity.MemberTitleType;
 import com.idea5.four_cut_photos_map.domain.memberTitle.repository.MemberTitleLogRepository;
 import com.idea5.four_cut_photos_map.domain.memberTitle.repository.MemberTitleRepository;
-import com.idea5.four_cut_photos_map.domain.review.service.ReviewService;
 import com.idea5.four_cut_photos_map.global.error.ErrorCode;
 import com.idea5.four_cut_photos_map.global.error.exception.BusinessException;
 import lombok.RequiredArgsConstructor;
@@ -28,9 +25,6 @@ import java.util.stream.Collectors;
 public class MemberTitleService {
     private final MemberTitleRepository memberTitleRepository;
     private final MemberTitleLogRepository memberTitleLogRepository;
-    private final FavoriteService favoriteService;
-
-    private final ReviewService reviewService;
 
     public MemberTitle findById(Long id) {
         return memberTitleRepository.findById(id).orElseThrow(() -> {
@@ -72,16 +66,6 @@ public class MemberTitleService {
         return memberTitleLogRepository.findByMember(member);
     }
 
-    @Transactional
-    public void addMemberTitle(Member member, Long memberTitleId, Boolean isMain) {
-        MemberTitle memberTitle = findById(memberTitleId);
-        memberTitleLogRepository.save(new MemberTitleLog(member, memberTitle, isMain));
-    }
-
-    public void addMemberTitle(Member member, MemberTitle memberTitle, Boolean isMain) {
-        memberTitleLogRepository.save(new MemberTitleLog(member, memberTitle, isMain));
-    }
-
     // 회원 대표 칭호 수정
     @Transactional
     public void updateMainMemberTitle(Member member, Long memberTitleId) {
@@ -118,35 +102,5 @@ public class MemberTitleService {
         return memberTitleLogRepository.findByMember(member)
                 .stream().map(memberTitleLog -> memberTitleLog.getMemberTitle().getId())
                 .collect(Collectors.toList());
-    }
-
-    // 회원에게 해당 회원칭호를 부여할 수 있는지 여부
-    public boolean canGiveMemberTitle(Member member, MemberTitle memberTitle) {
-        // TODO: 첫번째 리뷰, 리뷰 5개 이상 기준 추가
-        if(memberTitle.getId() == MemberTitleType.NEWBIE.getCode()) {
-            // 1. 회원가입
-            return true;
-        } else if(memberTitle.getId() == MemberTitleType.FIRST_REVIEW.getCode()) {
-            if(reviewService.getReviewCntByWriter(member) >= 1) {
-                // 2. 첫번째 리뷰
-                return true;
-            }
-        } else if(memberTitle.getId() == MemberTitleType.MANY_REVIEW.getCode()) {
-            if(reviewService.getReviewCntByWriter(member) >= 3) {
-                // 3. 리뷰 5개 이상
-                return true;
-            }
-        } else if(memberTitle.getId() == MemberTitleType.FIRST_HEART.getCode()) {
-            if(favoriteService.countByMember(member) >= 1) {
-                // 4. 첫번째 찜
-                return true;
-            }
-        } else if(memberTitle.getId() == MemberTitleType.MANY_HEART.getCode()) {
-            if(favoriteService.countByMember(member) >= 3) {
-                // 5. 찜 5개 이상
-                return true;
-            }
-        }
-        return false;
     }
 }
