+++ src/main/java/com/idea5/four_cut_photos_map/job/CollectJob.java
@@ -7,7 +7,6 @@ import com.idea5.four_cut_photos_map.domain.memberTitle.entity.MemberTitleType;
 import com.idea5.four_cut_photos_map.domain.memberTitle.service.MemberTitleService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StopWatch;
@@ -23,9 +22,10 @@ import java.util.List;
 public class CollectJob {
     private final MemberService memberService;
     private final MemberTitleService memberTitleService;
+    private final CollectService collectService;
 
     // 초 분 시 일 월 요일
-    @Scheduled(cron = "0 0/1 * * * *")      // TODO: 테스트용 매분마다 실행
+//    @Scheduled(cron = "0 * * * * *")      // TODO: 테스트용 매분마다 실행
 //    @Scheduled(cron = "0 0 0 * * *")    // 매일 0시 실행
     @Transactional
     public void add() {
@@ -50,10 +50,10 @@ public class CollectJob {
                 if(collectedMemberTitles.contains(memberTitle.getId()))
                     continue;
                 // 2. 회원이 보유하지 않은 회원칭호는 부여기준 검사 -> 부여
-                if(memberTitleService.canGiveMemberTitle(member, memberTitle)) {
+                if(collectService.canGiveMemberTitle(member, memberTitle)) {
                     // 회원가입 칭호와 다른 칭호를 같은 날에 부여 받는 경우 회원가입 칭호를 대표 칭호로 설정
                     boolean isMain = memberTitle.getId() == MemberTitleType.NEWBIE.getCode() ? true : false;
-                    memberTitleService.addMemberTitle(member, memberTitle, isMain);
+                    collectService.addMemberTitle(member, memberTitle, isMain);
                 }
             }
         }
